# PERF: Expand macros in parallel during CrateDefMap building