--- a/src/main/kotlin/org/rust/openapiext/utils.kt
+++ b/src/main/kotlin/org/rust/openapiext/utils.kt
@@ -22,9 +22,11 @@ import com.intellij.openapi.fileEditor.FileDocumentManager
 import com.intellij.openapi.fileEditor.impl.FileDocumentManagerImpl
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.module.ModuleManager
+import com.intellij.openapi.progress.EmptyProgressIndicator
 import com.intellij.openapi.progress.ProcessCanceledException
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.ProgressManager
+import com.intellij.openapi.progress.util.AbstractProgressIndicatorExBase
 import com.intellij.openapi.progress.util.ProgressIndicatorUtils
 import com.intellij.openapi.project.DumbService
 import com.intellij.openapi.project.IndexNotReadyException
@@ -36,6 +38,7 @@ import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.VirtualFileManager
 import com.intellij.openapi.vfs.VirtualFileWithId
+import com.intellij.openapi.wm.ex.ProgressIndicatorEx
 import com.intellij.psi.*
 import com.intellij.psi.impl.PsiDocumentManagerBase
 import com.intellij.psi.search.GlobalSearchScope
@@ -313,6 +316,24 @@ fun <T> executeUnderProgress(indicator: ProgressIndicator, action: () -> T): T {
     ProgressManager.getInstance().executeProcessUnderProgress({ result = action() }, indicator)
     @Suppress("UNCHECKED_CAST")
     return result ?: (null as T)
+}
+
+/**
+ * [this] indicator can be an instance of [com.intellij.openapi.progress.impl.BackgroundableProcessIndicator]
+ * class, which is thread sensitive and its [ProgressIndicator.checkCanceled] method should be used only from
+ * a single thread (see [com.intellij.openapi.progress.util.ProgressWindow.MyDelegate.checkCanceled]).
+ * So we propagate cancellation.
+ */
+fun ProgressIndicator.toThreadSafeProgressIndicator(): ProgressIndicator {
+    return if (this is ProgressIndicatorEx) {
+        val threadSafeIndicator = EmptyProgressIndicator()
+        addStateDelegate(object : AbstractProgressIndicatorExBase() {
+            override fun cancel() = threadSafeIndicator.cancel()
+        })
+        threadSafeIndicator
+    } else {
+        this
+    }
 }
 
 fun <T : PsiElement> T.createSmartPointer(): SmartPsiElementPointer<T> =