--- a/src/main/kotlin/org/rust/stdext/CompletableFuture.kt
+++ b/src/main/kotlin/org/rust/stdext/CompletableFuture.kt
@@ -8,11 +8,12 @@ package org.rust.stdext
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.ExecutionException
 import java.util.concurrent.Executor
+import java.util.concurrent.Future
 
 fun <T> supplyAsync(executor: Executor, supplier: () -> T): CompletableFuture<T> =
     CompletableFuture.supplyAsync({ supplier() }, executor)
 
-fun <T> CompletableFuture<T>.getWithRethrow(): T =
+fun <T> Future<T>.getWithRethrow(): T =
     try {
         get()
     } catch (e: ExecutionException) {