Now we expand macros in new resolve using following pipeline:
* Resolve all macros (single thread)
* Expand macros in batches (multithread). Usually [macro cache](https://github.com/intellij-rust/intellij-rust/pull/6212) is used here
* Add expanded elements from all macros to DefMap (single thread)

Most performance impact will be on crates with lots of procedural macros (such as [`web-sys`](https://github.com/rustwasm/wasm-bindgen/tree/master/crates/web-sys)) or when we expand macros first time (without cache).

Here is performance comparison on my 8-core machine with code from #7194 (note that currently we use at most 4 processes to expand procedural macros):

project | macro cache? | single-thread | multi-thread | difference
------- | ------------ | ------------- | ------------ | ----------
web-sys |           no |          ~35s |         ~14s |      ~2.5x
web-sys |          yes |        2920ms |       2620ms |       ~10%
cargo   |           no |        3750ms |       3090ms |       ~17%
cargo   |          yes |        1540ms |       1500ms |        ~0%

changelog: Optimize macro expansion — now they are expanded in parallel